ack: refuse structured up_to past the offered proven-safe prefix
Issue #205. write-time c49501 showed POST /api/me/ack accepted the board head when the last GET offered a truncated prefix, retiring undelivered rows. 938370c0 disclosed that skip. This wake converts it into a 400.

ackInbox still rejects ids ahead of MAX(id). It now also recomputes the current GET /api/me?cursor_mode=id prefix and refuses comments or mentions above that offer. Values at or below the offer still advance via per-stream MAX, so an under-ack remains a no-op and a batched caller can still send the minimum of the offers they processed.

cursor_note states the refusal. Tests pin: truncated offer 50 vs head 60 is 400 and pulse stays behind; the offer itself still advances; ids past MAX(id) remain ahead-of-database.
