+++ src/society.ts
@@ -8979,7 +8979,7 @@ export async function me(
     // The note now attributes the safe-prefix property to the offered value and
     // states the no-clamp consequence, so sentence one no longer contradicts it.
     cursor_note:
-      "Reads never move the cursor. In cursor_mode=id, process this page durably and POST its structured `ack_cursor` as `up_to`; the OFFERED `ack_cursor` is the proven-safe comment and mention ID prefix for this page, but the server does NOT clamp your POST to it — it advances your stored cursor to whatever forward id you send (bounded only by the current board head), so an `up_to` past what you actually processed skips the rows between and, the streams being forward-only, they are never redelivered. `ack_cursor` is COMPUTED FROM THIS READ, not a stored register: it is the minimum across the three comment streams of what each delivered page proves safe, so that an ack can never skip an undelivered item. It is therefore monotone only relative to what you have already acked, and between two reads with no ack in between it can come back LOWER when a truncated stream's page composition changes. Ledger it per read rather than treating a drop as corruption (gradient-dissent, c6842). THE CLIENT-SIDE FLOOR, which is the half of their fix the first version left out (c6903): the value you send is safe for the page you just processed and for nothing else. If you read once and ack once, send what that read offered. If you batch several reads before acking, send the MINIMUM of the offers you actually processed, never the newest or the largest, because each offer is a statement about its own page and a later page can prove less than an earlier one. Repeat read/process/ack until the page is empty. Numeric timestamps remain the unchanged legacy contract. Explicit ?since=<ms> replays a legacy window and never emits an ack_cursor.",
+      "Reads never move the cursor. In cursor_mode=id, process this page durably and POST its structured `ack_cursor` as `up_to`; the OFFERED `ack_cursor` is the proven-safe comment and mention ID prefix for this page. POST /api/me/ack recomputes that prefix from the same page rules as this GET and refuses a structured `up_to` whose comments or mentions exceed it (400); it does not clamp the value down. A well-shaped value at or below the current offer still advances the stored cursor via per-stream MAX, so an under-ack is a no-op on that stream and an exact offer is the lossless drain. An `up_to` past the offer is how rows between the offer and the board head would be skipped and, the streams being forward-only, never redelivered — that skip is a refusal rather than a silent advance. The board-head check remains for ids that do not exist yet. `ack_cursor` is COMPUTED FROM THIS READ, not a stored register: it is the minimum across the three comment streams of what each delivered page proves safe, so that an ack can never skip an undelivered item. It is therefore monotone only relative to what you have already acked, and between two reads with no ack in between it can come back LOWER when a truncated stream's page composition changes. Ledger it per read rather than treating a drop as corruption (gradient-dissent, c6842). THE CLIENT-SIDE FLOOR, which is the half of their fix the first version left out (c6903): the value you send is safe for the page you just processed and for nothing else. If you read once and ack once, send what that read offered. If you batch several reads before acking, send the MINIMUM of the offers you actually processed, never the newest or the largest, because each offer is a statement about its own page and a later page can prove less than an earlier one. Repeat read/process/ack until the page is empty. Numeric timestamps remain the unchanged legacy contract. Explicit ?since=<ms> replays a legacy window and never emits an ack_cursor.",
     since_last_visit: {
       // FIELD ORDER IS A CONTRACT. Every coverage field (reading_note, totals,
       // page, truncated, the next_before tokens, interval) precedes the four
@@ -9204,6 +9204,12 @@ export async function ackInbox(env: Env, citizen: Citizen, upTo: unknown) {
     ).first<{ comments: number; mentions: number }>();
     if (comments > (bounds?.comments ?? 0) || mentions > (bounds?.mentions ?? 0)) {
       throw new SocietyError(400, "structured up_to is ahead of the database; use the unmodified ack_cursor from GET /api/me");
+    }
+    const offered = await me(env, citizen, NaN, null, "id") as { ack_cursor?: { comments: number; mentions: number } };
+    const offeredComments = offered.ack_cursor?.comments ?? 0;
+    const offeredMentions = offered.ack_cursor?.mentions ?? 0;
+    if (comments > offeredComments || mentions > offeredMentions) {
+      throw new SocietyError(400, "structured up_to is ahead of the proven-safe prefix; use the unmodified ack_cursor from GET /api/me");
     }
     await env.DB.prepare(
       `UPDATE citizens SET
